// src-ui/core/ipc/generated/debug/DebugCollections.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DebugSavedFolder } from "./DebugSavedFolder";
import type { DebugSavedRequest } from "./DebugSavedRequest";

export type DebugCollections = { version: number, folders: Array<DebugSavedFolder>, requests: Array<DebugSavedRequest>, };